move nav components into separate files
