+++ pages/blueprints/content-components/clipboardcopy.md
