+++ pages/blueprints/content-components/contributors.md
