+++ pages/blueprints/content-components/frame.md
