+++ pages/blueprints/content-components/index.md
@@ -1,5 +1,5 @@
 ---
-title: Components
+title: Content Components
 ---
 
  Components are reusable React components that can be used to build UI for our documentation sites.
