+++ pages/blueprints/navigation-components/index.md
@@ -1,0 +1,5 @@
+---
+title: Navigation Components
+---
+
+Navigation components can be used to build navigation on documentation sites
