+++ src/components.js
@@ -5,8 +5,11 @@ export {default as Frame} from './Frame'
 export {default as Header} from './Header'
 export {default as Link} from './Link'
 export {default as NavLink} from './NavLink'
+export {default as NavList} from './NavList'
 export {default as NodeLink} from './NodeLink'
 export {default as Outline} from './Outline'
 export {default as PackageHeader} from './PackageHeader'
+export {default as RouteMatch} from './RouteMatch'
+export {default as Section} from './Section'
 export {default as SideNav} from './SideNav'
 export {default as StatusLabel} from './StatusLabel'
